refactor(runtime): split vm.ts into vm/ subdir, extract boot/snapshot helpers
Splits the 3795-LOC vm.ts into eight files under packages/runtime/src/vm/:
boot, attach, restore, snapshot, fork, bundle, image-config, helpers
(plus index.ts as the public barrel). Public surface unchanged —
@machinen/runtime consumers and provision.ts didn't need updates.

Also extracts named sub-functions from the two highest-cognitive
functions in the runtime:
  - boot() in vm/boot.ts: 242 -> 55 (16 helpers extracted)
  - performSnapshot() in vm/snapshot.ts: 77 -> off the targets list
    (6 helpers extracted)

Cleanup pass on findings flagged by fallow:
  - delete two unused example files (bake.ts, counter.mjs)
  - drop two dead exports from mount-server-detached.ts
  - remove two stale fallow-ignore suppressions in multiplex.ts
  - un-export MOUNT_ROOT and formatDumpOutcomeHint (file-internal)

The boot -> fork -> restore -> boot static cycle is left intact —
ESM tolerates it via function hoisting, and breaking it requires an
API change.
